(test) Add canonical chain discovery test
